accomadate $node->taxonomy data struct differences
When using search api, and using 'index heirarchy' of taxonomy terms,
the actual data structure in $node->whatever-taxonomy is different.
Because we index resources in its own index with 'index heirarchy' and
also as a general search index without 'index heirarchy' we must treat
taxonomy munging differently in each case.

Signed-off-by: Lee Walker <lee@codejourneymen.com>